requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# ----------------------
# API
# ----------------------
fastapi==0.123.4
uvicorn==0.38.0
pydantic==2.12.5
starlette==0.49.3
python-multipart==0.0.20

# ----------------------
# ML / NLP
# ----------------------
torch==2.2.2+cpu
transformers>=4.57.3,<5.0
tokenizers>=0.22,<0.24
safetensors>=0.4,<1.0
sentencepiece>=0.2
onnxruntime==1.19.2

gliner==0.2.24
huggingface-hub>=0.36,<1.0
hf-xet==1.2.0

# ----------------------
# Privacy
# ----------------------
presidio-analyzer==2.2.360
presidio-anonymizer==2.2.360

# ----------------------
# Utils
# ----------------------
numpy==1.26.4
regex==2024.5.15
tqdm==4.67.1

# ----------------------
# Documents / OCR
# ----------------------
pdfplumber==0.11.8
pdfminer.six==20251107
pytesseract==0.3.13
pillow==11.3.0
python-docx==1.2.0
